Counting points for keywords

+++ src/main/java/com/example/MovieTonight/dataFromApi/KeywordStatistic.java
@@ -8,7 +8,7 @@ import lombok.Setter;
 @Setter
 public class KeywordStatistic{
     private long genreId;
-    private int occurrences;
+    private long occurrences;
 
     public void increaseOccurrences(){
         this.occurrences++;
